backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Celery
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Email
sendgrid==6.11.0
jinja2==3.1.2

# File Processing
python-magic==0.4.27
pillow==10.1.0
reportlab==4.0.7
weasyprint==60.2

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Security Scanners Integration
python-owasp-zap-v2.4==0.0.21
nuclei-python==1.0.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0
schedule==1.2.0
dnspython==2.4.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0
prometheus-client==0.19.0